SamsungIntegration: Hide root functions from overflow menu

+++ app/src/main/java/projekt/substratum/adapters/fragments/themes/ThemeAdapter.java
@@ -190,20 +190,22 @@ public class ThemeAdapter extends RecyclerView.Adapter<ThemeAdapter.ViewHolder> 
                 });
 
         viewHolder.cardView.setOnLongClickListener(view -> {
-            // Vibrate the device alerting the user they are about to do something dangerous!
-            Vibrator v = (Vibrator) mContext.getSystemService(Context.VIBRATOR_SERVICE);
-            v.vibrate(30);
+            if (!References.isSamsung(mContext)) {
+                // Vibrate the device alerting the user they are about to do something dangerous!
+                Vibrator v = (Vibrator) mContext.getSystemService(Context.VIBRATOR_SERVICE);
+                v.vibrate(30);
 
-            SheetDialog sheetDialog = new SheetDialog(mContext);
-            View sheetView = View.inflate(mContext, R.layout.uninstall_sheet_dialog, null);
-            LinearLayout uninstall = sheetView.findViewById(R.id.uninstall);
-            uninstall.setOnClickListener(view2 -> {
-                toBeUninstalled = themeItem;
-                new uninstallTheme().execute();
-                sheetDialog.hide();
-            });
-            sheetDialog.setContentView(sheetView);
-            sheetDialog.show();
+                SheetDialog sheetDialog = new SheetDialog(mContext);
+                View sheetView = View.inflate(mContext, R.layout.uninstall_sheet_dialog, null);
+                LinearLayout uninstall = sheetView.findViewById(R.id.uninstall);
+                uninstall.setOnClickListener(view2 -> {
+                    toBeUninstalled = themeItem;
+                    new uninstallTheme().execute();
+                    sheetDialog.hide();
+                });
+                sheetDialog.setContentView(sheetView);
+                sheetDialog.show();
+            }
             return false;
         });
